requirements.txt
requests
yfinance
python-binance
pyTelegramBotAPI
websockets
ta
pandas
numpy
sqlalchemy
aiosqlite
lightgbm
scikit-learn
matplotlib
joblib